Add schemas for flows, validate data aginst them
## Feature

When someone hooks together flows to create a larger pipeline,
they need to know what data they can pass to the next flow.

## Solution

Create a `schema` field on each flow
that stores a [json-schema] definition for the flow's input.
When a flow is triggered through the web endpoint,
the input is evaluated against the schema
and validated before the script is run.

If the input is invalid,
the endpoint responds with `422 Unprocessable Entity`.
Otherwise, it runs the script normally and returns a `200 OK`.

[json-schema]: http://json-schema.org/examples.html

See:

* https://github.com/ruby-json-schema/json-schema
* https://spacetelescope.github.io/understanding-json-schema