Add user sign in with Monban
Monban is a gem that provides authentication with minimal opinions,
and supplies generators to support basic authentication flows.

## See

https://github.com/halogenandtoast/monban
https://github.com/halogenandtoast/monban-generators

## Notes

We ran into a problem setting up sign-in with email or username.
We opened up https://github.com/halogenandtoast/monban/pull/55
as a reproduction of the issue,
and with a potential fix.

We included that fix in our own application by overriding the default
Monban finder method in `config/initializers/monban.rb`.