Create basic apps to test building blocks
## Problem

Users need to test the building blocks that they create,
but building blocks can only be run in the context of an application.

## Solution

Add a link to program pages that invites users to "Try it out!".

When a user clicks the link,
the platform creates a new App
that contains a single block and is hooked up to a single entrypoint.

The entrypoint (still called a `Trigger`)
can accept events from an HTML form
that is displayed on the app page.

The app that is created is a special kind of app –
we call it a `SandboxApp`.
`SandboxApp`s are not yet represented in the database,
because they are currently
the only functional type of app on the platform.
If it makes sense,
we can break them out into their own model
using Single-Table Inheritance.
For now, we separate it from normal apps
by routing them through a different controller.
This lets us customize the interface to focus around the inputs and
outputs to the single app.

## Create connections within apps

  Connections can have a source and a destination.
  Sources can be Blocks or Triggers,
  and destinations can be Blocks or Triggers.

  Blocks are entirely contained within apps,
  while Triggers (soon to be renamed as EventStreams)
  are connections between an app and the outside world –
  both incoming and outgoing.

## Event Streams

- The `Event` model in the database represents events that start app and block runs.
- They store data in a `jsonb` column, which has some specific behavior:

  [Postgres JSON data type Rails query](http://stackoverflow.com/a/24463973/1857539)
  [Using PostgreSQL and jsonb with Ruby on Rails](http://nandovieira.com/using-postgresql-and-jsonb-with-ruby-on-rails)

## Notes

* Had to fix up a previous migration
  that got botched during a variable rename.
* We removed the direct connection between `Block`s and `Trigger`s.