Lazily generate block info for app canvas json
Previously, we stored detailed information about blocks
under `App#definition`.

This meant that once the block info was stored,
changes to the block would not be reflected in the app definition.

Even worse,
changes to the app canvas schema would completely invalidate apps,
and require that they be recreated from scratch.