Copy aaronshaf/react-toggle component into app
We copy this component directly,
instead of using it as an NPM module,
because we'd like to use styled-components instead of importing a plain
CSS file, as the package currently suggests.

In order to get the component working, we need to add a babel plugin to
support the `...spread` operator.

See https://babeljs.io/docs/plugins/transform-object-rest-spread/