Set up Clockwork gem to handle recurring events
Time-based triggers are an important part of getting Assemble off the ground.

While blocks on Assemble will ideally be based on real-time webhooks
powered by direct integrations with third-party services,
it will be time-intensive to develop the required integrations
for all the services we'd like.

Until we have create an integration for a service,
people can get by by scheduling recurring blocks to poll the service
and to take action when a change happens.

## Required Features

* Schedule a task to execute daily, weekly, or hourly
* Scheduled tasks should survive server restarts –
  this will likely require database persistence.

## Potential Libraries

* https://github.com/jmettraux/rufus-scheduler
* https://github.com/Rykian/clockwork
* https://github.com/treasure-data/perfectsched
* https://github.com/seejohnrun/ice_cube

## Attempt 1: Clockwork

https://github.com/Rykian/clockwork

Clockwork comes with a recommended way to set up database persistence.
It runs in a separate Ruby process,
which we'll likely want to spin up in a new Docker container.

First, we'll copy over the [example cloc.rb file][config] to `/util/clock.rb`.

[config]: https://github.com/Rykian/clockwork#example-clockrb-file

For persistence, Clockwork requires a database model
with a `#frequency` instance method.
We'll define that as `RecurringEvent`.

## Testing

We've set up a `RecurringEvent` in the database,
whose job it is to create an `Event` in the database.
However, this isn't happening.
The `sync` job works fine, but the events do not get created in the DB.

Aha! It turns out that `RecurringEvent#at`,
which was set to a default of "00:00",
was interfering with the normal scheduling.
Removing the default and setting that value to `nil`
gives us the behavior we expect.