Add `better_errors` gem for debugging
See https://github.com/charliesome/better_errors

I just ran into a problem that requires a debugger in a gem's code.
It's super hard to get a good breakpoint in there because of Docker,
so we need a new way to debug server errors.

`better_errors` really lives up to its name –
wheneer there's an error in our server code in development,
it gives us an entire browsable stack trace in-browser,
with a live REPL that can run at any level of the stack trace.