Introduce redux
## Problem

The app has accumulated a lot of client-side state
and a lot of logic for managing that state.
This leads to a few problems:

* Passing props down through several several levels of components
* Difficult to keep track of which components are responsible for
  different parts of application state
* Server update code is scattered throughout different components

## Solution

Redux gives us a single global state object for the web client,
and gives us a standard way to interact with that state through actions
and reducers.

This should help us lower the complexity of client state,
and keep it manageable over time.