Build custom client logic for API requests
## Problem

* We want full control over the shape of our Redux store
* Apollo doesn't integrate super well with Redux
* We'd like to co-locate GraphQL queries with the reducers that put them
in the store.

## Solution

* Create a new `requests` folder, with files that each store all the
  information for a single request.
* Create a helper to integrate all the info from a request file into the
  Redux / GraphQL workflows
* Integrate reducers to update the state with info from the response

## Hiccups

* Not able to use the `graphql-tag` loader, because after we load a
GraphQL query with it, it's not clear how to convert it back into a
plain GraphQL query string. It would be nice to figure this out, so we
can store queries in separate files, and use its fragment-fetching
feature.

## Future

There is still a lot of work to be done here –
there are many requests that are still being routed through Apollo.
From now on, we can gradually extract requests into our custom workflow,
using the `block` request as an example and a template.

We also have to sort out exactly how we want to organize the store,
and refactor out some higher-order reducers.