Move block environment info back onto block model
## Problem

Environment information is just as essential and individual to a block
as its source code is.

When we store environment information in a separate table,
we often need to touch two tables to change how a block works.

## Solution

Store block environment information alongside its source code,
on the block model itself.

This change removes functionality from the front-end –
it no longer allows people to change the environment of a block.

Now that we've got a stable & extensible system for updating state,
we should add a simple UI for editing blocks' environments.